refactor(service): update snils validator & user-docs service
